Now models cooling by water flow through pipes.
This is an initial water flow cooling model, and is not very sophisticated.  In order to not impact numerical stability stability the water flow portion of the model is not a dynamic model (dTwater/dt is assumed to be 0).  Also the isothermal condition of the concrete layers becomes less realistic once cooling through a pipe is involved, as this would in reality set up temperature gradients.