Modification of CCT1D version 3.1
While this extends from version 4, it goes back to the cooling architecture of version 3.1 wherein cooling is applied to some layers but not others.  It models cooling water flow rates, inlet temperatures, pipe sizes and materials.  There are some relatively minor enhancements to the output data read out and to how the output data file is name.